% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_best_records}
\alias{load_best_records}
\title{Load Best Records Information}
\usage{
load_best_records(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads best records CSV from Samsung Health Data.
}
