% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_blood_pressure}
\alias{load_blood_pressure}
\title{Load Blood Pressure}
\usage{
load_blood_pressure(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Blood Pressure CSV from Samsung Health Data.
}
