% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_day_summary}
\alias{load_day_summary}
\title{Load Day Summary Information}
\usage{
load_day_summary(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads day summary CSV from Samsung Health Data.
   Note: does not load extra data yet.
}
