% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_device_profile}
\alias{load_device_profile}
\title{Load Device Profile}
\usage{
load_device_profile(path)
}
\arguments{
\item{path}{path to the CSV file}
}
\description{
Load a device profile CSV from Samsung Health Data.
}
