% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_food_frequent}
\alias{load_food_frequent}
\title{Load Frequent Food}
\usage{
load_food_frequent(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Frequent Food CSV from Samsung Health Data.
}
