% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_food_info}
\alias{load_food_info}
\title{Load Food Information}
\usage{
load_food_info(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads food information CSV from Samsung Health Data.
}
