% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_goal_history}
\alias{load_goal_history}
\title{Load Goal History}
\usage{
load_goal_history(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Goal History CSV from Samsung Health Data.
}
