% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_hourly_accumulated_active_time}
\alias{load_hourly_accumulated_active_time}
\title{Load Hourly Accumulated Active Time}
\usage{
load_hourly_accumulated_active_time(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Hourly Accumulated Active Time CSV from Samsung Health Data.
}
