% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_leaderboard}
\alias{load_leaderboard}
\title{Load leaderboard}
\usage{
load_leaderboard(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads leaderboard CSV from Samsung Health Data.
}
