% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_oxygen_saturation}
\alias{load_oxygen_saturation}
\title{Load Oxygen Saturation}
\usage{
load_oxygen_saturation(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Oxygen Saturation CSV from Samsung Health Data.
}
