% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_pedometer_day_summary}
\alias{load_pedometer_day_summary}
\title{Load Daily Pedometer Summary}
\usage{
load_pedometer_day_summary(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Daily Pedometer Summary CSV from Samsung Health Data.
}
