% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_pedometer_event}
\alias{load_pedometer_event}
\title{Load Pedometer Events}
\usage{
load_pedometer_event(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Pedometer Events CSV from Samsung Health Data.
}
