% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_pedometer_step_count}
\alias{load_pedometer_step_count}
\title{Load Pedometer Step Count}
\usage{
load_pedometer_step_count(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Pedometer Step Count CSV from Samsung Health Data.
}
