% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_preferences}
\alias{load_preferences}
\title{Load preferences}
\usage{
load_preferences(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads preferences CSV from Samsung Health Data.
}
