% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_samsunghealth}
\alias{load_samsunghealth}
\title{Load Samsung Health Data}
\usage{
load_samsunghealth(path)
}
\arguments{
\item{path}{Path to the exported Samsung Health Data}
}
\description{
load_samsunghealth loads a Samsung Health Data export
}
