% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_service_status}
\alias{load_service_status}
\title{Load Service Status}
\usage{
load_service_status(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Service Status CSV from Samsung Health Data.
}
