% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_step_daily_trend}
\alias{load_step_daily_trend}
\title{Load Daily Step Trend}
\usage{
load_step_daily_trend(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads Daily Step Trend CSV from Samsung Health Data.
}
