% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_user_profile}
\alias{load_user_profile}
\title{Load User Profile Information}
\usage{
load_user_profile(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads user profile CSV from Samsung Health Data.
   Note: does not load dashboard config yet.
}
