% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_weight}
\alias{load_weight}
\title{Load Weight Information}
\usage{
load_weight(path)
}
\arguments{
\item{path}{path to CSV file}
}
\description{
Loads weight CSV from Samsung Health Data. Note that
   Samsung reports weight in kilograms, even if the data was input
   in Pounds.
}
